BEOperations/requirements.txt
aiohttp==3.5.4
alabaster==0.7.12
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-classy-tags==0.9.0
django-cms==3.6.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.1
django-model-utils==3.2.0
django-multiselectfield==0.1.8
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sekizai==1.0.0
django-tables2==2.1.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-snippet==2.2.0
djangorestframework==3.10.2
djangorestframework-datatables==0.5.0
djangorestframework-sso==0.1.0
docutils==0.15.2
enum-compat==0.0.2
eventlet==0.21.0
Flask==1.1.1
geos==0.2.2
gis==0.2.1
greenlet==0.4.15
httpie==1.0.2
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
more-itertools==7.2.0
multidict==4.5.2
numpy==1.16.4
openapi-codec==1.3.2
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
postgis==1.0.4
psycopg2==2.8.3
py==1.8.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.0.1
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
websockets==8.0.2
Werkzeug==0.15.4
yarl==1.3.0
zipp==0.5.2
